+++ app/views/tutorial/_prerequisites.html.erb
@@ -1,13 +1,13 @@
 <h1>Configure your prerequisites</h1>
 
 <p>
-  There are a few prerequisites that you need to complete before you can work through this task. If you've already completed any of them, feel free to skip that step
+  There are a few prerequisites that you need to complete before you can work through this tutorial. If you've already completed any of them, feel free to skip that step
 </p>
 
 <hr />
 
 
-<% @task.prerequisites.map do |p| %>
+<% @tutorial.prerequisites.map do |p| %>
   <div class="Vlt-accordion">
     <button class="Vlt-accordion__trigger" style="padding-top: 10px;"><h4 style="margin-bottom:10px;"><%= p['title'] %></h4></button>
