+++ config/tutorials/sms-fortune-cookies-with-ibm-cloud.yml
